[VAS-1119] refactor to unify response
