+++ src/main/java/it/pagopa/selfcare/pagopa/backoffice/model/creditorinstituions/CreditorInstitutionInfo.java
@@ -1,6 +1,5 @@
 package it.pagopa.selfcare.pagopa.backoffice.model.creditorinstituions;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -23,7 +22,6 @@ public class CreditorInstitutionInfo {
     @NotNull
     private String businessName;
 
-    @JsonProperty("creditorInstitutionCode")
     @Schema(example = "02438750586", description = "The tax code of the creditor institution", requiredMode = Schema.RequiredMode.REQUIRED)
     @NotBlank
     private String ciTaxCode;
